improved basic route handling
